Fix requests failing on cloud
