+++ src/plugins/console/server/lib/utils/index.ts
@@ -7,3 +7,4 @@
  */
 
 export { encodePath } from './encode_path';
+export { toURL } from './to_url';
