+++ src/plugins/console/server/routes/api/console/proxy/query_string.test.ts
@@ -39,11 +39,11 @@ describe('Console Proxy Route', () => {
   describe('query string', () => {
     describe('path', () => {
       describe('contains full url', () => {
-        it.skip('treats the url as a path', async () => {
+        it('treats the url as a path', async () => {
           await request('GET', 'http://evil.com/test');
           expect(proxyRequestMock.mock.calls.length).toBe(1);
           const [[args]] = (requestModule.proxyRequest as jest.Mock).mock.calls;
-          expect(args.uri.href).toBe('http://localhost:9200/http://evil.com/test?pretty=true');
+          expect(args.uri.href).toBe('http://localhost:9200/http%3A//evil.com/test?pretty=true');
         });
       });
       describe('starts with a slash', () => {
